+++ 1415_2TIB4_aventure/levelgesture.cpp
@@ -1,7 +1,7 @@
 #include "flib.h"
 #include "levelgesture.h"
 
-const int TILE = 32;		// Tiles size
+const float TILE = 32;		// Tiles size
 
 const int XSCREENLENGHT = 15;	// Number of tiles on x
